remove unnecessary floating point casts in examples

+++ examples/bayesian_linear_regression_tensor.py
@@ -24,8 +24,7 @@ def build_toy_dataset(N, noise_std=0.1):
   X = np.concatenate([np.linspace(0, 2, num=N / 2),
                       np.linspace(6, 8, num=N / 2)])
   y = 5.0 * X + np.random.normal(0, noise_std, size=N)
-  X = X.astype(np.float32).reshape((N, 1))
-  y = y.astype(np.float32)
+  X = X.reshape((N, 1))
   return X, y
 
 
@@ -38,7 +37,7 @@ D = 1  # num features
 X_data, y_data = build_toy_dataset(N)
 
 # MODEL
-X = X_data
+X = tf.cast(X_data, tf.float32)
 w = Normal(mu=tf.zeros(D), sigma=tf.ones(D))
 b = Normal(mu=tf.zeros(1), sigma=tf.ones(1))
 y = Normal(mu=ed.dot(X, w) + b, sigma=tf.ones(N))
